moves abstracts, functions, and modesl to utils entry point
