+++ libs/crispr-forms/public-api.ts
@@ -1,2 +1,3 @@
 export * from './src/index';
+// eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
 export * from '@tft/crispr-forms/utils';
