+++ libs/crispr-forms/src/lib/material/autocomplete-field/autocomplete-field.component.ts
@@ -1,8 +1,6 @@
 import { Component, OnInit, ViewChild, ChangeDetectionStrategy, ElementRef, OnDestroy } from '@angular/core';
 import { MatAutocompleteTrigger, MatAutocompleteSelectedEvent, MatAutocompleteModule } from '@angular/material/autocomplete';
 
-import { SelectOption, AutocompleteFieldConfig, DEFAULT_EMPTY_OPTIONS_MESSAGE } from '../../models';
-import { AbstractAutocompleteComponent } from '../../abstracts';
 import { Subscription } from 'rxjs';
 import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
@@ -11,6 +9,8 @@ import { InfoComponent } from '../info/info.component';
 import { MatInputModule } from '@angular/material/input';
 import { MatOptionModule } from '@angular/material/core';
 import { OptionComponent } from '../option';
+// eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
+import { SelectOption, AbstractAutocompleteComponent, AutocompleteFieldConfig, DEFAULT_EMPTY_OPTIONS_MESSAGE } from '@tft/crispr-forms/utils';
 
 const defaultConfig: Partial<AutocompleteFieldConfig> = {
   autoActiveFirstOption: true,
