+++ libs/crispr-forms/src/lib/material/option/option.component.ts
@@ -1,7 +1,8 @@
 import { CommonModule } from '@angular/common';
 import { Component, Input, ChangeDetectionStrategy } from '@angular/core';
-import { SelectOption } from '../../models';
 import { InfoComponent } from '../info/info.component';
+// eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
+import { SelectOption } from '@tft/crispr-forms/utils';
 
 @Component({
   selector: 'crispr-option',
