+++ libs/crispr-forms/src/lib/material/slider-field/slider-field.component.ts
@@ -1,12 +1,12 @@
 import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
-import { SliderFieldConfig } from '../../models';
-import { CrisprFieldComponent, crisprControlMixin } from '../../abstracts';
 import { MatSliderModule } from '@angular/material/slider';
 import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
 import { FieldContainerComponent } from '../../field-container';
 import { InfoComponent } from '../info/info.component';
 import { OptionComponent } from '../option';
+// eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
+import { SliderFieldConfig, CrisprFieldComponent, crisprControlMixin } from '@tft/crispr-forms/utils';
 
 const defaultConfig = {displayLimits: true};
 const SliderFieldMixin = crisprControlMixin<SliderFieldConfig>(CrisprFieldComponent);
