+++ libs/crispr-forms/src/lib/material/textarea-field/textarea-field.component.ts
@@ -1,12 +1,12 @@
 import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
-import { TextareaFieldConfig } from '../../models';
-import { crisprControlMixin, CrisprFieldComponent } from '../../abstracts';
 import { CommonModule } from '@angular/common';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { FieldContainerComponent } from '../../field-container';
 import { InfoComponent } from '../info/info.component';
+// eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
+import { crisprControlMixin, CrisprFieldComponent, TextareaFieldConfig } from '@tft/crispr-forms/utils';
 
 const defaultConfig: Partial<TextareaFieldConfig> = {rows: 5};
 const TextareaFieldMixin = crisprControlMixin<TextareaFieldConfig>(CrisprFieldComponent);
