+++ libs/crispr-forms/utils/models/image-upload-field.config.ts
