+++ libs/crispr-forms/utils/models/slider-field.config.ts
